Reviewed-on: https://git.onlyoffice.com/ONLYOFFICE/onlyoffice-docspace-pipedrive/pulls/22

+++ frontend/src/pages/Settings/Authorization.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Command } from "@pipedrive/app-extensions-sdk";
+import { Command, View } from "@pipedrive/app-extensions-sdk";
 import { DocSpace } from "@onlyoffice/docspace-react";
 import { TFrameConfig } from "@onlyoffice/docspace-sdk-js/dist/types/types";
 import { SDKInstance } from "@onlyoffice/docspace-sdk-js/dist/types/instance";
@@ -16,11 +16,19 @@ import { putDocspaceAccount, deleteDocspaceAccount } from "@services/user";
 
 import Authorized from "@assets/authorized.svg";
 import NotAvailable from "@assets/not-available.svg";
+import Welcome from "@assets/welcome.svg";
+
 import { ErrorResponse } from "src/types/error";
 
 const DOCSPACE_SYSTEM_FRAME_ID = "authorization-docspace-system-frame";
 
-export const AuthorizationSetting: React.FC = () => {
+export type AuthorizationSettingProps = {
+  showUserGuide(): void;
+};
+
+export const AuthorizationSetting: React.FC<AuthorizationSettingProps> = ({
+  showUserGuide,
+}) => {
   const { t } = useTranslation();
   const {
     user,
@@ -34,28 +42,68 @@ export const AuthorizationSetting: React.FC = () => {
 
   const [saving, setSaving] = useState(false);
   const [deleting, setDeleting] = useState(false);
-  const [showValidationMessage, setShowValidationMessage] = useState(false);
+
   const [email, setEmail] = useState<string | undefined>("");
+  const [isInvalidEmail, setIsInvalidEmail] = useState(false);
+  const [errorTextInvalidEmail, setErrorTextInvalidEmail] = useState("");
+
   const [password, setPassword] = useState<string | undefined>("");
+  const [isInvalidPassword, setIsInvalidPassword] = useState(false);
+  const [errorTextInvalidPassword, setErrorTextInvalidPassword] = useState("");
 
   let docspaceInstance: SDKInstance;
 
   const handleLogin = async (event: React.SyntheticEvent) => {
     event.preventDefault();
-    if (email && password) {
-      setSaving(true);
-    } else {
-      setShowValidationMessage(true);
+
+    if (!email || !password) {
+      if (!email) {
+        setIsInvalidEmail(true);
+        setErrorTextInvalidEmail(
+          t("error.empty-field", "This field is required"),
+        );
+      } else {
+        setIsInvalidEmail(false);
+      }
+
+      if (!password) {
+        setIsInvalidPassword(true);
+        setErrorTextInvalidPassword(
+          t("error.empty-field", "This field is required"),
+        );
+      } else {
+        setIsInvalidPassword(false);
+      }
+
+      return;
     }
+
+    setIsInvalidEmail(false);
+    setIsInvalidPassword(false);
+
+    setSaving(true);
   };
 
   const handleLogout = async () => {
+    const { confirmed } = await sdk.execute(Command.SHOW_CONFIRMATION, {
+      title: t("label.warning", "Warning"),
+      description:
+        t(
+          "settings.authorization.deleting.confirm-message",
+          "Are you sure you want to log out?",
+        ) || "",
+      okText: t("button.logout", "Log out"),
+    });
+
+    if (!confirmed) {
+      return;
+    }
+
     setDeleting(true);
     deleteDocspaceAccount(pipedriveToken)
       .then(async () => {
         setEmail("");
         setPassword("");
-        setShowValidationMessage(false);
         if (user && settings) {
           setUser({ ...user, docspaceAccount: null });
         }
@@ -99,8 +147,16 @@ export const AuthorizationSetting: React.FC = () => {
       clearTimeout(loginTimeout);
 
       if (login.status && login.status !== 200) {
+        setIsInvalidEmail(true);
+        setErrorTextInvalidEmail(
+          t("docspace.error.login", "User authentication failed"),
+        );
+
         await sdk.execute(Command.SHOW_SNACKBAR, {
-          message: t("docspace.error.login", "User authentication failed"),
+          message: t(
+            "settings.authorization.unsuccessful",
+            "Invalid authorization credentials. Please check your Email and password and try to log in again.",
+          ),
         });
         setSaving(false);
       } else {
@@ -126,6 +182,7 @@ export const AuthorizationSetting: React.FC = () => {
                 "ONLYOFFICE DocSpace authorization has been successfully saved",
               ),
             });
+            showUserGuide();
           })
           .catch(async (e) => {
             const data = e?.response?.data as ErrorResponse;
@@ -261,7 +318,7 @@ export const AuthorizationSetting: React.FC = () => {
               <OnlyofficeTitle
                 text={t(
                   "settings.authorization.title",
-                  "Login to ONLYOFFICE DocSpace account",
+                  "Log into your connected ONLYOFFICE DocSpace to start using it within Pipedrive",
                 )}
               />
             </div>
@@ -277,29 +334,6 @@ export const AuthorizationSetting: React.FC = () => {
               </div>
             )}
           </div>
-          {user?.docspaceAccount && (
-            <>
-              <div className="inline-flex pl-5 pr-5">
-                <div className="p-1">
-                  <Authorized />
-                </div>
-                <span className="pl-3">
-                  {t(
-                    "settings.authorization.status.authorized",
-                    "You have successfully logged in to your ONLYOFFICE DocSpace account",
-                  )}
-                </span>
-              </div>
-              <div className="flex justify-start items-center mt-4 ml-5">
-                <OnlyofficeButton
-                  text={t("button.logout", "Log out")}
-                  color={ButtonColor.PRIMARY}
-                  loading={deleting}
-                  onClick={handleLogout}
-                />
-              </div>
-            </>
-          )}
           {!user?.docspaceAccount && (
             <div className="max-w-[390px]">
               <form onSubmit={handleLogin}>
@@ -310,7 +344,9 @@ export const AuthorizationSetting: React.FC = () => {
                       "settings.authorization.inputs.email",
                       "Email",
                     )}
-                    valid={showValidationMessage ? !!email : true}
+                    required
+                    valid={!isInvalidEmail}
+                    errorText={errorTextInvalidEmail}
                     value={email}
                     disabled={saving}
                     onChange={(e) => setEmail(e.target.value.trim())}
@@ -326,16 +362,22 @@ export const AuthorizationSetting: React.FC = () => {
                       "settings.authorization.inputs.password",
                       "Password",
                     )}
-                    valid={showValidationMessage ? !!password : true}
+                    required
+                    valid={!isInvalidPassword}
+                    errorText={errorTextInvalidPassword}
                     value={password}
                     type="password"
                     disabled={saving}
+                    link={{
+                      text: t("button.forgot-password", "Forgot password?"),
+                      href: `${settings.url}/profile`,
+                    }}
                     onChange={(e) => setPassword(e.target.value)}
                   />
                 </div>
                 <div className="flex justify-start items-center mt-4 ml-5">
                   <OnlyofficeButton
-                    text={t("button.login", "Login")}
+                    text={t("button.login", "Log in")}
                     type="submit"
                     color={ButtonColor.PRIMARY}
                     loading={saving}
@@ -344,6 +386,68 @@ export const AuthorizationSetting: React.FC = () => {
                 </div>
               </form>
             </div>
+          )}
+          {user?.docspaceAccount && (
+            <>
+              <div className="flex gap-3 mt-1 pb-2 pl-5 pr-5">
+                <div>
+                  <Authorized />
+                </div>
+                <div className="flex justify-center items-center">
+                  <p>
+                    {t(
+                      "settings.authorization.status.authorized",
+                      "You have successfully logged in to your ONLYOFFICE DocSpace account",
+                    )}{" "}
+                    <span className="font-semibold text-pipedrive-color-light-green-600 dark:text-pipedrive-color-dark-green-600">
+                      {user.docspaceAccount.userName}
+                    </span>
+                  </p>
+                </div>
+              </div>
+              <div className="flex justify-start items-center mt-4 ml-5">
+                <OnlyofficeButton
+                  text={t("button.logout", "Log out")}
+                  color={ButtonColor.NEGATIVE}
+                  loading={deleting}
+                  onClick={handleLogout}
+                />
+              </div>
+              <div className="pt-4">
+                <OnlyofficeBackgroundError
+                  Icon={<Welcome />}
+                  title={t(
+                    "settings.authorization.welcome.title",
+                    "Welcome to DocSpace!",
+                  )}
+                  options={[
+                    t(
+                      "settings.authorization.welcome.option1",
+                      "Create a room in the deal",
+                    ),
+                    t(
+                      "settings.authorization.welcome.option2",
+                      "Upload a document to the room",
+                    ),
+                  ]}
+                  button={{
+                    text: t("button.deals", "Go to Deals"),
+                    onClick: async () => {
+                      await sdk.execute(Command.REDIRECT_TO, {
+                        view: View.DEALS,
+                      });
+                    },
+                  }}
+                  link={{
+                    text: t(
+                      "settings.authorization.welcome.open-guide",
+                      "Open Guide",
+                    ),
+                    onClick: showUserGuide,
+                  }}
+                />
+              </div>
+            </>
           )}
         </>
       )}
